requirements.txt
# requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
praw>=7.5.0
nltk>=3.6.5
textblob>=0.17.1
pandas>=1.3.3
python-dotenv>=0.19.1
spacy>=3.2.0
transformers>=4.11.3
torch>=1.9.1